修复搜索页面显示高亮bug和用户个人主页头像显示bug

+++ routes/web.php
@@ -21,25 +21,26 @@ Route::get('/', function () {
     return redirect(route('topic.index'));
 });
 
+Route::get('/user-banned','UserController@isBanned');
 
 
 Route::group(['middleware' => ['web']], function () {
     Route::get('search','SearchController@index')->name('search');
     Route::group(['as' => 'topic.'], function () {
-        Route::get('topic/create','TopicController@create')->name('create')->middleware('auth');
-        Route::get('topic/{id}/edit','TopicController@edit')->name('edit')->middleware('auth');
+        Route::get('topic/create','TopicController@create')->name('create')->middleware(['auth','banned_auth']);
+        Route::get('topic/{id}/edit','TopicController@edit')->name('edit')->middleware(['auth','banned_auth']);
         Route::get('topic/{id}','TopicController@show')->name('show');
         Route::get('topic','TopicController@index')->name('index');
-        Route::match(['put', 'patch'],'topic/{id}/edit','TopicController@update')->name('update')->middleware('auth');
-        Route::post('topic/uploadimg','TopicController@uploadImg')->name('uploadImg')->middleware('auth');
-        Route::post('topic/vote','TopicController@vote')->name('vote')->middleware('auth');
-        Route::post('topic','TopicController@store')->name('store')->middleware('auth');
-        Route::delete('topic/{id}','TopicController@destroy')->name('delete')->middleware('auth');
+        Route::match(['put', 'patch'],'topic/{id}/edit','TopicController@update')->name('update')->middleware(['auth','banned_auth']);
+        Route::post('topic/uploadimg','TopicController@uploadImg')->name('uploadImg')->middleware(['auth','banned_auth']);
+        Route::post('topic/vote','TopicController@vote')->name('vote')->middleware(['auth','banned_auth']);
+        Route::post('topic','TopicController@store')->name('store')->middleware(['auth','banned_auth']);
+        Route::delete('topic/{id}','TopicController@destroy')->name('delete')->middleware(['auth','banned_auth']);
     });
     Route::get('cate/{id}','CategoryController@show')->name('category.show');
     Route::get('tag/{id}','TagController@show')->name('tag.show');
-    Route::post('reply','ReplyController@store')->name('reply.create')->middleware('auth');
-    Route::delete('reply/{id}','ReplyController@destroy')->name('reply.delete')->middleware('auth');
+    Route::post('reply','ReplyController@store')->name('reply.create')->middleware(['auth','banned_auth']);
+    Route::delete('reply/{id}','ReplyController@destroy')->name('reply.delete')->middleware(['auth','banned_auth']);
    // Route::get('user/messages','UserController@getMessages')->name('user.getMessages');
     Route::get('user/notifications','UserController@getNotifications')->name('user.notifications');
     Route::get('user/mentions','UserController@getMentions')->name('user.mentions');
@@ -49,20 +50,20 @@ Route::group(['middleware' => ['web']], function () {
     Route::get('user/{id}/fans','UserController@getFansList')->name('user.fansList');
     Route::get('user/{id}/votes','UserController@getVotesList')->name('user.votesList');
     Route::get('user/{id}/excellence','UserController@getExcellencesList')->name('user.excellencesList');
-    Route::get('user/{id}/edit','UserController@edit')->name('user.edit')->middleware('auth');
-    Route::get('user/{id}/portrait','UserController@editPortrait')->name('user.editPortrait')->middleware('auth');
-    Route::get('user/{id}/password','UserController@editPassword')->name('user.editPassword')->middleware('auth');
+    Route::get('user/{id}/edit','UserController@edit')->name('user.edit')->middleware(['auth','banned_auth']);
+    Route::get('user/{id}/portrait','UserController@editPortrait')->name('user.editPortrait')->middleware(['auth','banned_auth']);
+    Route::get('user/{id}/password','UserController@editPassword')->name('user.editPassword')->middleware(['auth','banned_auth']);
     Route::get('user/{id}','UserController@show')->name('user.show');
-    Route::post('follower','UserController@follower')->name('follower.create')->middleware('auth');
-    Route::delete('follower/{id}','UserController@followerDestroy')->name('follower.delete')->middleware('auth');
-    Route::match(['put', 'patch'],'user/{id}/portrait','UserController@updatePortrait')->name('user.updatePortrait')->middleware('auth');
-    Route::match(['put', 'patch'],'user/{id}/password','UserController@updatePassword')->name('user.updatePassword')->middleware('auth');
-    Route::match(['put', 'patch'],'user/{id}','UserController@update')->name('user.update')->middleware('auth');
+    Route::post('follower','UserController@follower')->name('follower.create')->middleware(['auth','banned_auth']);
+    Route::delete('follower/{id}','UserController@followerDestroy')->name('follower.delete')->middleware(['auth','banned_auth']);
+    Route::match(['put', 'patch'],'user/{id}/portrait','UserController@updatePortrait')->name('user.updatePortrait')->middleware(['auth','banned_auth']);
+    Route::match(['put', 'patch'],'user/{id}/password','UserController@updatePassword')->name('user.updatePassword')->middleware(['auth','banned_auth']);
+    Route::match(['put', 'patch'],'user/{id}','UserController@update')->name('user.update')->middleware(['auth','banned_auth']);
 
 
 
 });
 
-Route::group(['prefix' => 'admin', 'middleware' => ['auth', 'admin_auth']], function () {
+Route::group(['prefix' => 'admin', 'middleware' => ['auth','banned_auth', 'admin_auth']], function () {
     Route::get('{path?}', 'AdminController@index')->where('path', '[\/\w\.-]*')->name('admin.index');
 });
